feature/#1 ajout dossier compenent et modif des vues login et register
